backend/tools/flights_amadeus.py
# backend/tools/flights_amadeus.py
from amadeus import Client, ResponseError
from tenacity import retry, stop_after_attempt, wait_exponential
from backend.config import settings

def _client() -> Client:
    if not settings.AMADEUS_CLIENT_ID or not settings.AMADEUS_CLIENT_SECRET:
        raise RuntimeError("Amadeus credentials missing")
    return Client(
        client_id=settings.AMADEUS_CLIENT_ID,
        client_secret=settings.AMADEUS_CLIENT_SECRET
    )

@retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=0.5, min=0.5, max=2))
def flight_eur(origin: str, dest_airport: str, depart_date: str) -> dict:
    """
    Uses Amadeus Flight Offers Search (IATA codes recommended).
    """
    am = _client()
    try:
        res = am.shopping.flight_offers_search.get(
            originLocationCode=origin,
            destinationLocationCode=dest_airport,
            departureDate=depart_date,
            adults=1,
            currencyCode="EUR",
            max=5,
        )
        data = res.data or []
        price = min([float(o["price"]["total"]) for o in data]) if data else 140.0
    except ResponseError:
        price = 140.0

    return {
        "provider": "amadeus",
        "price_eur": round(price, 2),
        "currency": "EUR",
        "url": "https://developers.amadeus.com/",
        "ttl_min": 15,
    }