requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
orjson==3.10.7
redis==5.0.8
psycopg[binary]==3.2.1
SQLAlchemy==2.0.36
alembic==1.13.2
qdrant-client==1.11.3
sentence-transformers==3.0.1
numpy==1.26.4
pandas==2.2.2
geopy==2.4.1
pytest==8.3.2
pytest-asyncio==0.23.8
streamlit==1.38.0
langchain==0.2.14
langgraph==0.2.15

